Lifting state up to Game.
